backend/requirement.txt
fastapi==0.127.0
uvicorn==0.40.0
pymongo==4.15.5
pydantic==2.12.5
email-validator==2.3.0
python-dotenv==1.2.1
python-multipart==0.0.21

